requirements.txt
beir==2.2.0
faiss-cpu==1.13.2
jupyter==1.1.1
notebook==7.3.2
numpy==2.4.3
pandas==3.0.1
pytrec_eval-terrier==0.5.10
rank-bm25==0.2.2
sentence-transformers==5.3.0
torch==2.10.0
